+++ com.unity.render-pipelines.universal/Runtime/Debug/UniversalRenderPipelineDebugDisplaySettings.cs
@@ -9,40 +9,19 @@ namespace UnityEngine.Rendering.Universal
         /// <summary>
         /// Material-related Rendering Debugger settings.
         /// </summary>
-        internal DebugDisplaySettingsMaterial MaterialSettings { get; private set; }
+        public DebugDisplaySettingsMaterial materialSettings { get; private set; }
 
         /// <summary>
         /// Rendering-related Rendering Debugger settings.
         /// </summary>
-        internal DebugDisplaySettingsRendering RenderingSettings { get; private set; }
+        public DebugDisplaySettingsRendering renderingSettings { get; private set; }
 
         /// <summary>
         /// Lighting-related Rendering Debugger settings.
         /// </summary>
-        internal DebugDisplaySettingsLighting LightingSettings { get; private set; }
+        public DebugDisplaySettingsLighting lightingSettings { get; private set; }
 
         #region IDebugDisplaySettingsQuery
-
-        /// <summary>
-        /// Returns true if any of the debug settings are currently active.
-        /// </summary>
-        public override bool AreAnySettingsActive => MaterialSettings.AreAnySettingsActive ||
-        LightingSettings.AreAnySettingsActive ||
-        RenderingSettings.AreAnySettingsActive;
-
-        public override bool TryGetScreenClearColor(ref Color color)
-        {
-            return MaterialSettings.TryGetScreenClearColor(ref color) ||
-                RenderingSettings.TryGetScreenClearColor(ref color) ||
-                LightingSettings.TryGetScreenClearColor(ref color);
-        }
-
-        /// <summary>
-        /// Returns true if lighting is active for current state of debug settings.
-        /// </summary>
-        public override bool IsLightingActive => MaterialSettings.IsLightingActive &&
-        RenderingSettings.IsLightingActive &&
-        LightingSettings.IsLightingActive;
 
         /// <summary>
         /// Returns true if the current state of debug settings allows post-processing.
@@ -51,7 +30,7 @@ namespace UnityEngine.Rendering.Universal
         {
             get
             {
-                DebugPostProcessingMode debugPostProcessingMode = RenderingSettings.debugPostProcessingMode;
+                DebugPostProcessingMode debugPostProcessingMode = renderingSettings.postProcessingDebugMode;
 
                 switch (debugPostProcessingMode)
                 {
@@ -62,10 +41,11 @@ namespace UnityEngine.Rendering.Universal
 
                     case DebugPostProcessingMode.Auto:
                     {
-                        // Only enable post-processing if we aren't using certain debug-views...
-                        return MaterialSettings.IsPostProcessingAllowed &&
-                            RenderingSettings.IsPostProcessingAllowed &&
-                            LightingSettings.IsPostProcessingAllowed;
+                        // Only enable post-processing if we aren't using certain debug-views.
+                        bool postProcessingAllowed = true;
+                        foreach (IDebugDisplaySettingsData setting in m_Settings)
+                            postProcessingAllowed &= setting.IsPostProcessingAllowed;
+                        return postProcessingAllowed;
                     }
 
                     case DebugPostProcessingMode.Enabled:
@@ -84,6 +64,7 @@ namespace UnityEngine.Rendering.Universal
 
         public UniversalRenderPipelineDebugDisplaySettings()
         {
+            Reset();
         }
 
         public override void Reset()
@@ -91,9 +72,9 @@ namespace UnityEngine.Rendering.Universal
             m_Settings.Clear();
 
             CommonSettings = Add(new DebugDisplaySettingsCommon());
-            MaterialSettings = Add(new DebugDisplaySettingsMaterial());
-            LightingSettings = Add(new DebugDisplaySettingsLighting());
-            RenderingSettings = Add(new DebugDisplaySettingsRendering());
+            materialSettings = Add(new DebugDisplaySettingsMaterial());
+            lightingSettings = Add(new DebugDisplaySettingsLighting());
+            renderingSettings = Add(new DebugDisplaySettingsRendering());
         }
     }
 }